Move fnmatch.c/h to misc
